feat: add ERC5564Announcer contract and tests (#2)
See [ERC-5564](https://eips.ethereum.org/EIPS/eip-5564) for the spec. The contract is super simple so we really only have one test, if you have ideas for more tests let me know.

This PR also turns off the metadata hash, that way the resulting bytecode is independent of the machine it's compiled on.